+++ sdk/keyvault/Azure.Security.KeyVault.Certificates/tests/CertificateClientLiveTests.cs
@@ -19,6 +19,19 @@ namespace Azure.Security.KeyVault.Certificates.Tests
         {
         }
 
+        [SetUp]
+        public void ClearChallengeCacheforRecord()
+        {
+            // In record mode we reset the challenge cache before each test so that the challenge call
+            // is always made. This allows tests to be replayed independently and in any order
+            if (Mode == RecordedTestMode.Record || Mode == RecordedTestMode.Playback)
+            {
+                Client = GetClient();
+
+                ChallengeBasedAuthenticationPolicy.AuthenticationChallenge.ClearCache();
+            }
+        }
+
         [Test]
         public void StartCreateCertificateError()
         {