+++ sdk/keyvault/Azure.Security.KeyVault.Certificates/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Fixed
 
+- Reset challenge cache so tests can be run individually and in any order. (#9356)[https://github.com/Azure/azure-sdk-for-net/issues/9356]
 - Properly import PKCS12 or PEM-encoded certificate. (#9963)[https://github.com/Azure/azure-sdk-for-net/issues/9963]
 
 ## 4.0.0 (2019-01-08)